Bughunt on kmer, sort.  Tracing user-reported bugs on Shred, MapPacBio.

+++ current/bloom/KCountArray3.java
@@ -84,7 +84,7 @@ public class KCountArray3 extends KCountArray {
 		int value=((word>>>cellShift)&valueMask);
 		if(value==0 && incr>0){cellsUsed++;}
 		else if(incr<0 && value+incr==0){cellsUsed--;}
-		value=min(value+incr, maxValue);
+		value=(int)min(value+(long)incr, maxValue);
 		word=(value<<cellShift)|(word&~((valueMask)<<cellShift));
 		array[index]=word;
 		if(verbose){System.err.println("Returning "+value);}
@@ -109,7 +109,7 @@ public class KCountArray3 extends KCountArray {
 		int word=array[index];
 		int cellShift=(int)(cellBits*key);
 		final int value=((word>>>cellShift)&valueMask);
-		final int value2=min(value+incr, maxValue);
+		final int value2=(int)min(value+(long)incr, maxValue);
 		word=(value2<<cellShift)|(word&~((valueMask)<<cellShift));
 		array[index]=word;
 		if(verbose){System.err.println("Returning "+value);}
